# docs/swagger.yaml
basePath: /api
definitions:
...
  model.Pipeline:
    type: object
  model.PipelineStep:
    type: object
  model.PromptTemplate:
    type: object
  resp.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
host: localhost:7847
info:
  contact: {}
  description: LLM 文档处理与知识图谱 API 中间件
  title: Nexus API
  version: "1.0"
paths:
...
  /admin/pipelines:
    get:
      description: 获取所有 Pipeline
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 获取 Pipeline 列表
      tags:
      - Pipeline
    post:
      consumes:
      - application/json
      description: 创建新的处理流水线
      parameters:
      - description: Pipeline 配置
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.Pipeline'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 创建 Pipeline
      tags:
      - Pipeline
  /admin/pipelines/{id}:
    delete:
      description: 删除指定 Pipeline
      parameters:
      - description: Pipeline ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 删除 Pipeline
      tags:
      - Pipeline
    get:
      description: 获取指定 Pipeline 及其步骤
      parameters:
      - description: Pipeline ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 获取 Pipeline 详情
      tags:
      - Pipeline
    put:
      consumes:
      - application/json
      description: 更新 Pipeline 配置
      parameters:
      - description: Pipeline ID
        in: path
        name: id
        required: true
        type: integer
      - description: Pipeline 配置
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.Pipeline'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 更新 Pipeline
      tags:
      - Pipeline
  /admin/pipelines/{id}/steps:
    post:
      consumes:
      - application/json
      description: 为指定 Pipeline 添加处理步骤
      parameters:
      - description: Pipeline ID
        in: path
        name: id
        required: true
        type: integer
      - description: 步骤配置
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.PipelineStep'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 创建 Pipeline 步骤
      tags:
      - Pipeline
  /admin/pipelines/{id}/steps/{step_id}:
    delete:
      description: 删除指定步骤
      parameters:
      - description: Pipeline ID
        in: path
        name: id
        required: true
        type: integer
      - description: 步骤 ID
        in: path
        name: step_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 删除 Pipeline 步骤
      tags:
      - Pipeline
    put:
      consumes:
      - application/json
      description: 更新指定步骤配置
      parameters:
      - description: Pipeline ID
        in: path
        name: id
        required: true
        type: integer
      - description: 步骤 ID
        in: path
        name: step_id
        required: true
        type: integer
      - description: 步骤配置
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.PipelineStep'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 更新 Pipeline 步骤
      tags:
      - Pipeline
  /admin/pipelines/{id}/steps/reorder:
    put:
      consumes:
      - application/json
      description: 调整步骤执行顺序
      parameters:
      - description: Pipeline ID
        in: path
        name: id
        required: true
        type: integer
      - description: '步骤 ID 列表 {step_ids: [1,2,3]}'
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 重排 Pipeline 步骤
      tags:
      - Pipeline
  /admin/prompts:
    get:
      description: 获取所有提示词模板
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 获取提示词模板列表
      tags:
      - 提示词
    post:
      consumes:
      - application/json
      description: 创建新的提示词模板
      parameters:
      - description: 模板内容
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.PromptTemplate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 创建提示词模板
      tags:
      - 提示词
  /admin/prompts/{id}:
    delete:
      description: 删除指定提示词模板
      parameters:
      - description: 模板 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 删除提示词模板
      tags:
      - 提示词
    get:
      description: 获取指定提示词模板
      parameters:
      - description: 模板 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 获取提示词模板详情
      tags:
      - 提示词
    put:
      consumes:
      - application/json
      description: 更新指定提示词模板
      parameters:
      - description: 模板 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 模板内容
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.PromptTemplate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 更新提示词模板
      tags:
      - 提示词
  /admin/reviews:
    get:
      description: 分页获取审核记录，支持按状态筛选
      parameters:
      - description: 状态筛选
        in: query
        name: status
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 获取审核列表
      tags:
      - 审核
  /admin/reviews/{id}/approve:
    put:
      description: 通过指定审核记录
      parameters:
      - description: 审核 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 通过审核
      tags:
      - 审核
  /admin/reviews/{id}/modify:
    put:
      consumes:
      - application/json
      description: 修改审核记录的提取结果
      parameters:
      - description: 审核 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 修改内容
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 修改审核内容
      tags:
      - 审核
  /admin/reviews/{id}/reject:
    put:
      description: 拒绝指定审核记录
      parameters:
      - description: 审核 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 拒绝审核
      tags:
      - 审核
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"